Got navbar and a register button
The navbar is not functional yet, but the button is in the right place and it looks right, so it just needs to have routing added to that button. The register component needs to be styled so that the user can actually type in details and then the component should grab that data and output it to the console.